Add dependencies to the model.
Right now all dependencies are passed as jars now even
for jars that are actually Gradle sub-projects.

+++ builder/src/main/java/com/android/builder/DefaultManifestParser.java
@@ -71,7 +71,7 @@ public class DefaultManifestParser implements ManifestParser {
         try {
             Object value = AndroidManifest.getMinSdkVersion(new FileWrapper(manifestFile));
             if (value instanceof Integer) {
-                return ((Integer)value).intValue();
+                return (Integer) value;
             } else if (value instanceof String) {
                 // TODO: support codename
             }
